Tests moved to packages
