+++ .gitignore
@@ -29,4 +29,4 @@ _testmain.go
 *.prof
 
 # Go dependency workspace
-Godeps/_workspace/*
+vendor/*
