+++ Godeps/Godeps.json
@@ -2,9 +2,13 @@
 	"ImportPath": "github.com/nopecmd/nope",
 	"GoVersion": "go1.6",
 	"Packages": [
-		"github.com/nopecmd/nope/cmd/nope/"
+		"./..."
 	],
 	"Deps": [
+		{
+			"ImportPath": "github.com/davecgh/go-spew/spew",
+			"Rev": "5215b55f46b2b919f50a1df0eaa5886afe4e3b3d"
+		},
 		{
 			"ImportPath": "github.com/flynn/go-shlex",
 			"Rev": "3f9db97f856818214da2e1057f8ad84803971cff"
@@ -19,8 +23,17 @@
 			"Rev": "e6c6068076470196af082b1ff896e24a51a87b2a"
 		},
 		{
+			"ImportPath": "github.com/pmezard/go-difflib/difflib",
+			"Rev": "792786c7400a136282c1664665ae0a8db921c6c2"
+		},
+		{
 			"ImportPath": "github.com/rogpeppe/rog-go/reverse",
 			"Rev": "f57ad5e24ab7813942c045f86eda1066eb969e98"
+		},
+		{
+			"ImportPath": "github.com/stretchr/testify/assert",
+			"Comment": "v1.1.3-6-g6fe211e",
+			"Rev": "6fe211e493929a8aac0469b93f28b1d0688a9a3a"
 		}
 	]
 }
