--- test/util_test.go
@@ -1,5 +1,0 @@
-package nope_test
-
-func formatError(rawCmd string, msg string) string {
-	return rawCmd + " command failed: " + msg
-}
