+++ utils/utils.go
@@ -1,0 +1,5 @@
+package utils
+
+func FormatError(rawCmd string, msg string) string {
+	return rawCmd + " command failed: " + msg
+}
